@@ -2,7 +2,7 @@
 name = "spec-runner"
 version = "0.6.1"
 edition = "2021"
-rust-version = "1.66.0"
+rust-version = "1.67.0"
 license = "MIT"
 authors = ["Ryan Lopopolo <rjl@hyperbo.la>"]
 description = "Binary for running Ruby Specs with Artichoke Ruby"
