@@ -70,8 +70,7 @@ mod tests {
         for path in test_cases {
             assert!(
                 !is_explicit_relative(OsStr::new(path)),
-                "expected absolute path '{}' to NOT be explicit relative path",
-                path
+                "expected absolute path '{path}' to NOT be explicit relative path"
             );
         }
     }
@@ -82,8 +81,7 @@ mod tests {
         for path in test_cases {
             assert!(
                 !is_explicit_relative(OsStr::new(path)),
-                "expected relative path '{}' to NOT be explicit relative path",
-                path
+                "expected relative path '{path}' to NOT be explicit relative path"
             );
         }
     }
@@ -94,8 +92,7 @@ mod tests {
         for path in test_cases {
             assert!(
                 is_explicit_relative(OsStr::new(path)),
-                "expected relative path '{}' to be explicit relative path",
-                path
+                "expected relative path '{path}' to be explicit relative path"
             );
         }
     }
@@ -106,8 +103,7 @@ mod tests {
         for path in test_cases {
             assert!(
                 !is_explicit_relative(OsStr::new(path)),
-                "expected path '{}' to NOT be explicit relative path",
-                path
+                "expected path '{path}' to NOT be explicit relative path"
             );
         }
     }
@@ -127,8 +123,7 @@ mod tests {
         for path in test_cases {
             assert!(
                 !is_explicit_relative(OsStr::new(path)),
-                "expected relative path '{}' to NOT be explicit relative path",
-                path
+                "expected relative path '{path}' to NOT be explicit relative path"
             );
         }
     }
