@@ -104,8 +104,7 @@ mod tests {
         for path in test_cases {
             assert!(
                 !is_explicit_relative(OsStr::new(path)),
-                "expected absolute path '{}' to NOT be explicit relative path",
-                path
+                "expected absolute path '{path}' to NOT be explicit relative path"
             );
         }
     }
@@ -116,8 +115,7 @@ mod tests {
         for path in test_cases {
             assert!(
                 !is_explicit_relative(OsStr::new(path)),
-                "expected relative path '{}' to NOT be explicit relative path",
-                path
+                "expected relative path '{path}' to NOT be explicit relative path"
             );
         }
     }
@@ -137,8 +135,7 @@ mod tests {
         for path in test_cases {
             assert!(
                 is_explicit_relative(OsStr::new(path)),
-                "expected relative path '{}' to be explicit relative path",
-                path
+                "expected relative path '{path}' to be explicit relative path"
             );
         }
     }
@@ -149,8 +146,7 @@ mod tests {
         for path in test_cases {
             assert!(
                 !is_explicit_relative(OsStr::new(path)),
-                "expected path '{}' to NOT be explicit relative path",
-                path
+                "expected path '{path}' to NOT be explicit relative path"
             );
         }
     }
